Cut `value` into pieces
This is for great support
Alteration
Of piecing
Don't have the luck to tidy diffs up, reading
