+++ src/value/ser.rs
@@ -1,0 +1,26 @@
+use serde::{self, Serialize};
+
+use super::Value;
+
+impl Serialize for Value {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+        where S: serde::Serializer
+    {
+        match *self {
+            Value::Null => serializer.serialize_unit(),
+            Value::Bool(b) => serializer.serialize_bool(b),
+            Value::I64(i) => serializer.serialize_i64(i),
+            Value::F64(f) => serializer.serialize_f64(f),
+            Value::String(ref s) => serializer.serialize_str(s),
+            Value::Sequence(ref seq) => seq.serialize(serializer),
+            Value::Mapping(ref hash) => {
+                use serde::ser::SerializeMap;
+                let mut map = serializer.serialize_map(Some(hash.len()))?;
+                for (k, v) in hash {
+                    map.serialize_entry(k, v)?;
+                }
+                map.end()
+            }
+        }
+    }
+}
